+++ plugin_base/src/plugin_base.clap/plugin_base.clap/pb_plugin.hpp
@@ -25,6 +25,7 @@ public any_state_listener,
 public gui_param_listener
 {
   typedef moodycamel::ReaderWriterQueue<sync_event, default_q_size> event_queue;
+  typedef moodycamel::ReaderWriterQueue<mod_indicator_state, default_q_size> mod_indicator_queue;
 
   // needs to be first, everyone else needs it
   std::unique_ptr<plugin_desc> _desc;
@@ -36,6 +37,8 @@ public gui_param_listener
   std::vector<int> _block_automation_seen = {};
   std::unique_ptr<event_queue> _to_gui_events = {};
   std::unique_ptr<event_queue> _to_audio_events = {};
+  std::unique_ptr<mod_indicator_queue> _mod_indicator_queue = {};
+  std::vector<mod_indicator_state> _mod_indicator_states = {};
 
   // see param_state_changed and timerCallback()
   // and vst3 pb_controller _inside_set_param_normalized