+++ plugin_base/src/plugin_base/plugin_base/dsp/block/host.cpp
@@ -7,10 +7,11 @@ void
 host_block::prepare()
 {
   audio_out = nullptr;
-  events.out.clear();
+  events.midi.clear();
   events.notes.clear();
   events.block.clear();
-  events.midi.clear();
+  events.output_params.clear();
+  events.mod_indicator_states.clear();
   events.accurate_automation.clear();
   events.accurate_modulation.clear();
   events.accurate_automation_and_modulation.clear();
@@ -23,17 +24,18 @@ host_block::prepare()
 void 
 host_events::deactivate()
 {
-  notes = {};
-  out = {};
-  block = {};
   midi = {};
+  notes = {};
+  block = {};
+  output_params = {};
+  mod_indicator_states = {};
   accurate_automation = {};
   accurate_modulation = {};
   accurate_automation_and_modulation = {};
 }
 
 void 
-host_events::activate(bool graph, int param_count, int midi_count, int polyphony, int max_frame_count)
+host_events::activate(bool graph, int module_count, int param_count, int midi_count, int polyphony, int max_frame_count)
 { 
   // reserve this much but allocate on the audio thread if necessary
   // still seems better than dropping events
@@ -43,11 +45,16 @@ host_events::activate(bool graph, int param_count, int midi_count, int polyphony
   int note_limit_guess = polyphony * fill_guess;
   int midi_events_guess = midi_count * fill_guess;
   int accurate_events_guess = param_count * fill_guess;
+  int mod_indicator_states_guess = module_count * polyphony;
 
-  notes.reserve(note_limit_guess);
-  out.reserve(block_events_guess);
-  block.reserve(block_events_guess);
   midi.reserve(midi_events_guess);
+  notes.reserve(note_limit_guess);
+  block.reserve(block_events_guess);
+  output_params.reserve(block_events_guess);
+
+  // see also plugin_engine::ctor
+  mod_indicator_states.reserve(mod_indicator_states_guess);
+
   accurate_automation.reserve(accurate_events_guess);
   accurate_modulation.reserve(accurate_events_guess);
   accurate_automation_and_modulation.reserve(accurate_events_guess);