+++ plugin_base/src/plugin_base/plugin_base/dsp/block/shared.hpp
@@ -17,4 +17,18 @@ struct shared_block final {
   float const* const* audio_in;
 };
 
+struct mod_indicator_state_data final {
+  std::uint8_t module;
+  std::uint8_t module_slot;
+  std::uint8_t voice;
+  std::uint8_t user;
+  float value;
+};
+
+// visual modulation indication
+union mod_indicator_state final {
+  std::uint64_t packed;
+  mod_indicator_state_data data;
+};
+
 }