+++ plugin_base/src/plugin_base/plugin_base/dsp/splice_engine.cpp
@@ -113,7 +113,9 @@ plugin_splice_engine::activate(int max_frame_count)
   else _splice_block_size = 128;
 
   _engine.activate(_splice_block_size);
-  _host_block.events.activate(false, state().desc().param_count, state().desc().midi_count, state().desc().plugin->audio_polyphony, max_frame_count);
+  _host_block.events.activate(false, 
+    state().desc().module_count, state().desc().param_count, 
+    state().desc().midi_count, state().desc().plugin->audio_polyphony, max_frame_count);
 
   // make some room for the block boundary / interpolation events
   int fill_guess = (int)std::ceil(max_frame_count / 32.0f);
@@ -133,7 +135,9 @@ plugin_splice_engine::process()
   assert(spliced_block_count * spliced_block_frames + rest_block_frames == _host_block.frame_count);
   int total_block_count = spliced_block_count + (rest_block_frames == 0? 0: 1);
 
-  _host_block.events.out.clear();
+  _host_block.events.output_params.clear();
+  _host_block.events.mod_indicator_states.clear();
+
   splice_accurate_events(
     _host_block.events.accurate_automation, _spliced_accurate_automation_events, 
     _host_block.frame_count, spliced_block_count, spliced_block_frames, rest_block_frames);
@@ -212,8 +216,22 @@ plugin_splice_engine::process()
         inner_block.events.accurate_modulation.push_back(e);
       }
 
-    _engine.process();
-    _host_block.events.out.insert(_host_block.events.out.begin(), inner_block.events.out.begin(), inner_block.events.out.end());
+    _engine.process();  
+
+    _host_block.events.output_params.insert(
+      _host_block.events.output_params.begin(),
+      inner_block.events.output_params.begin(),
+      inner_block.events.output_params.end());
+
+    // only care about the last one
+    if (i == total_block_count - 1)
+    {
+      _host_block.events.mod_indicator_states.insert(
+        _host_block.events.mod_indicator_states.begin(),
+        inner_block.events.mod_indicator_states.begin(),
+        inner_block.events.mod_indicator_states.end());
+    }
+
     _engine.release_block();
   }
 }