+++ plugin_base/src/plugin_base/plugin_base/gui/graph.cpp
@@ -11,6 +11,7 @@ module_graph::
 { 
   _done = true;
   stopTimer();
+  _gui->remove_mod_indicator_state_listener(this);
   if(_module_params.render_on_tweak) _gui->gui_state()->remove_any_listener(this);
   if(_module_params.render_on_tab_change) _gui->remove_tab_selection_listener(this);
   if (_module_params.render_on_module_mouse_enter || _module_params.render_on_param_mouse_enter_modules.size())
@@ -33,6 +34,7 @@ graph(lnf, params), _gui(gui), _module_params(module_params)
     gui->add_tab_selection_listener(this);
     module_tab_changed(_module_params.module_index, 0);
   }
+  _gui->add_mod_indicator_state_listener(this);
   startTimerHz(_module_params.fps);
 }
 
@@ -49,6 +51,64 @@ module_graph::timerCallback()
   if(_done || !_render_dirty) return;
   if(render_if_dirty())
     repaint();
+}
+
+void 
+module_graph::mod_indicator_state_changed(std::vector<mod_indicator_state> const& states)
+{
+  for (int i = 0; i < max_indicators; i++)
+    _indicators[i]->setVisible(false);
+
+  if (_data.type() != graph_data_type::series)
+    return;
+
+  if (_hovered_or_tweaked_param == -1)
+    return;
+
+  float w = getWidth();
+  float h = getHeight();
+  int count = _data.series().size();
+
+  int current_indicator = 0;
+  int current_module_slot = -1;
+  int current_module_index = -1;
+
+  auto const& topo = *_gui->gui_state()->desc().plugin;
+  auto const& mappings = _gui->gui_state()->desc().param_mappings.params;
+  param_topo_mapping mapping = mappings[_hovered_or_tweaked_param].topo;
+
+  if (_module_params.module_index != -1)
+  {
+    current_module_slot = _activated_module_slot;
+    current_module_index = _module_params.module_index;
+  }
+  else
+  {
+    current_module_slot = _gui->gui_state()->desc().param_mappings.params[_hovered_or_tweaked_param].topo.module_slot;
+    current_module_index = _gui->gui_state()->desc().param_mappings.params[_hovered_or_tweaked_param].topo.module_index;
+  }
+
+  if (topo.modules[current_module_index].mod_indicator_source_selector != nullptr)
+  {
+    auto selected = topo.modules[current_module_index].mod_indicator_source_selector(*_gui->gui_state(), mapping);
+    if (selected.module_index != -1 && selected.module_slot != -1)
+    {
+      current_module_slot = selected.module_slot;
+      current_module_index = selected.module_index;
+    }
+  }
+
+  for (int i = 0; i < states.size() && current_indicator < max_indicators; i++, current_indicator++)
+    if (current_module_index == states[i].data.module && current_module_slot == states[i].data.module_slot)
+    {
+      float indicator_pos = states[i].data.value;      
+      float x = indicator_pos * w;
+      int point = std::clamp((int)(indicator_pos * (count - 1)), 0, count - 1);
+      float y = (1 - std::clamp(_data.series()[point], 0.0f, 1.0f)) * h;
+      _indicators[current_indicator]->setVisible(true);
+      _indicators[current_indicator]->setBounds(x - 3, y - 3, 6, 6);
+      _indicators[current_indicator]->repaint();
+    }
 }
 
 void 
@@ -139,12 +199,39 @@ module_graph::render_if_dirty()
   if (_hovered_or_tweaked_param == -1) return false;
 
   auto const& mappings = _gui->gui_state()->desc().param_mappings.params;
-  param_topo_mapping mapping = mapping = mappings[_hovered_or_tweaked_param].topo;
+  param_topo_mapping mapping = mappings[_hovered_or_tweaked_param].topo;
   auto const& module = _gui->gui_state()->desc().plugin->modules[mapping.module_index];
   if(module.graph_renderer != nullptr)
-    render(module.graph_renderer(*_gui->gui_state(), _gui->get_module_graph_engine(module), _hovered_or_tweaked_param, mapping));
+    render(module.graph_renderer(
+      *_gui->gui_state(), _gui->get_module_graph_engine(module), _hovered_or_tweaked_param, mapping));
   _render_dirty = false;
   return true;
+}
+
+graph_indicator::
+graph_indicator(lnf* lnf) : _lnf(lnf)
+{
+  setSize(6, 6);
+  setVisible(false);
+}
+
+void
+graph_indicator::paint(Graphics& g)
+{
+  g.setColour(_lnf->colors().graph_mod_indicator);
+  g.fillEllipse(0, 0, 6, 6);
+}
+
+graph::
+graph(lnf* lnf, graph_params const& params) :
+  _lnf(lnf), _data(graph_data_type::na, {}), _params(params)
+{
+  _indicators.resize(max_indicators);
+  for (int i = 0; i < max_indicators; i++)
+  {
+    _indicators[i] = std::make_unique<graph_indicator>(_lnf);
+    addChildComponent(_indicators[i].get());
+  }
 }
 
 void