+++ plugin_base/src/plugin_base/plugin_base/shared/graph_data.cpp
@@ -16,8 +16,8 @@ graph_data::init(graph_data const& rhs)
   case graph_data_type::off: break;
   case graph_data_type::scalar: _scalar = rhs.scalar(); break;
   case graph_data_type::audio: _audio = jarray<float, 2>(rhs.audio()); break;
-  case graph_data_type::series: _series = jarray<float, 1>(rhs.series()); break;
   case graph_data_type::multi_stereo: _multi_stereo = rhs.multi_stereo(); break;
+  case graph_data_type::series: _series = jarray<float, 1>(rhs.series()); break;
   default: assert(false);
   }
 }