+++ plugin_base/src/plugin_base/plugin_base/topo/module.hpp
@@ -7,6 +7,7 @@
 #include <plugin_base/shared/jarray.hpp>
 #include <plugin_base/shared/utility.hpp>
 #include <plugin_base/shared/graph_data.hpp>
+#include <plugin_base/dsp/block/shared.hpp>
 
 #include <memory>
 #include <vector>
@@ -26,6 +27,12 @@ class state_converter;
 
 enum class module_output { none, cv, audio };
 enum class module_stage { input, voice, output };
+
+struct mod_indicator_source
+{
+  int module_index;
+  int module_slot;
+};
 
 class module_tab_menu_result {
   bool _show_warning = false;
@@ -81,8 +88,14 @@ typedef std::function<std::unique_ptr<module_engine>(
   plugin_topo const& topo, int sample_rate, int max_frame_count)> 
 module_engine_factory;
 typedef std::function<graph_data(
-  plugin_state const& state, graph_engine* engine, int param, param_topo_mapping const& mapping)>
+  plugin_state const& state, graph_engine* engine, 
+  int param, param_topo_mapping const& mapping)>
 module_graph_renderer;
+
+// in case we want to plot someone elses mod indicators
+typedef std::function<mod_indicator_source(
+  plugin_state const& state, param_topo_mapping const& mapping)>
+  module_mod_indicator_source_selector;
 
 // module topo mapping
 struct module_topo_mapping final {
@@ -166,14 +179,15 @@ struct module_topo final {
   std::vector<midi_source> midi_sources;
   midi_active_selector midi_active_selector_;
 
+  state_initializer minimal_initializer;
+  state_initializer default_initializer;
+  bool force_rerender_on_param_hover = false;
+
   module_graph_renderer graph_renderer;
   module_engine_factory engine_factory;
   module_graph_engine_factory graph_engine_factory;
   module_state_converter_factory state_converter_factory;
-
-  state_initializer minimal_initializer;
-  state_initializer default_initializer;
-  bool force_rerender_on_param_hover = false;
+  module_mod_indicator_source_selector mod_indicator_source_selector;
 
   PB_PREVENT_ACCIDENTAL_COPY_DEFAULT_CTOR(module_topo);
   void validate(plugin_topo const& plugin, int index) const;