+++ src/firefly_synth/firefly_synth/envelope.cpp
@@ -101,6 +101,7 @@ private:
   cv_filter _filter = {};
 
   double _stage_pos = 0;
+  double _total_pos = 0;
   double _current_level = 0;
   double _multitrig_level = 0;
 
@@ -137,20 +138,16 @@ init_default(plugin_state& state)
 { state.set_text_at(module_env, 1, param_on, 0, "On"); }
 
 void
-env_plot_length_seconds(plugin_state const& state, int slot, float& dahds, float& dahdsrf)
+env_plot_length_seconds(plugin_state const& state, int slot, float& dly, float& att, float& hld, float& dcy, float& rls, float& flt)
 {
   float const bpm = 120;
   bool sync = state.get_plain_at(module_env, slot, param_sync, 0).step() != 0;
-  int type = state.get_plain_at(module_env, slot, param_type, 0).step();
-  float filter = state.get_plain_at(module_env, slot, param_filter, 0).real() / 1000.0f;
-  float hold = sync_or_time_from_state(state, bpm, sync, module_env, slot, param_hold_time, param_hold_tempo);
-  float delay = sync_or_time_from_state(state, bpm, sync, module_env, slot, param_delay_time, param_delay_tempo);
-  float decay = sync_or_time_from_state(state, bpm, sync, module_env, slot, param_decay_time, param_decay_tempo);
-  float attack = sync_or_time_from_state(state, bpm, sync, module_env, slot, param_attack_time, param_attack_tempo);
-  float release = sync_or_time_from_state(state, bpm, sync, module_env, slot, param_release_time, param_release_tempo);
-  float sustain = type != type_sustain? 0.0f : std::max((delay + attack + hold + decay + release + filter) / 5, 0.01f);
-  dahds = delay + attack + hold + decay + sustain;
-  dahdsrf = dahds + release + filter;
+  flt = state.get_plain_at(module_env, slot, param_filter, 0).real() / 1000.0f;
+  hld = sync_or_time_from_state(state, bpm, sync, module_env, slot, param_hold_time, param_hold_tempo);
+  dly = sync_or_time_from_state(state, bpm, sync, module_env, slot, param_delay_time, param_delay_tempo);
+  dcy = sync_or_time_from_state(state, bpm, sync, module_env, slot, param_decay_time, param_decay_tempo);
+  att = sync_or_time_from_state(state, bpm, sync, module_env, slot, param_attack_time, param_attack_tempo);
+  rls = sync_or_time_from_state(state, bpm, sync, module_env, slot, param_release_time, param_release_tempo);
 }
 
 static graph_engine_params
@@ -158,7 +155,7 @@ make_graph_engine_params()
 {
   graph_engine_params result = {};
   result.bpm = 120;
-  result.max_frame_count = 200;
+  result.max_frame_count = 2000; // need to spot outliers
   return result;
 }
 
@@ -170,23 +167,26 @@ make_graph_engine(plugin_desc const* desc)
 }
 
 static graph_data
-render_graph(plugin_state const& state, graph_engine* engine, int param, param_topo_mapping const& mapping)
+render_graph(
+  plugin_state const& state, graph_engine* engine, int param, param_topo_mapping const& mapping)
 {
   if (state.get_plain_at(module_env, mapping.module_slot, param_on, 0).step() == 0) 
     return graph_data(graph_data_type::off, {});
 
-  float dahds;
-  float dahdsrf;
-  env_plot_length_seconds(state, mapping.module_slot, dahds, dahdsrf);
-  if(dahdsrf < 1e-5) return graph_data({}, false, {"0 Sec"});
+  float dly, att, hld, dcy, rls, flt;
+  env_plot_length_seconds(state, mapping.module_slot, dly, att, hld, dcy, rls, flt);
+  float dahd = dly + att + hld + dcy;
+  float dahdrf = dahd + rls + flt;
+
+  if(dahdrf < 1e-5) return graph_data({}, false, {"0 Sec"});
   
-  std::string partition = float_to_string(dahdsrf, 2) + " Sec";
+  std::string partition = float_to_string(dahdrf, 2) + " Sec";
   bool sync = state.get_plain_at(module_env, mapping.module_slot, param_sync, 0).step() != 0;
-  if(sync) partition = float_to_string(dahdsrf / timesig_to_time(120, { 1, 1 }), 2) + " Bar";
+  if(sync) partition = float_to_string(dahdrf / timesig_to_time(120, { 1, 1 }), 2) + " Bar";
 
   auto const params = make_graph_engine_params();
-  int sample_rate = params.max_frame_count / dahdsrf;
-  int voice_release_at = dahds / dahdsrf * params.max_frame_count;
+  int sample_rate = params.max_frame_count / dahdrf;
+  int voice_release_at = dahd / dahdrf * params.max_frame_count;
 
   engine->process_begin(&state, sample_rate, params.max_frame_count, voice_release_at);
   auto const* block = engine->process(module_env, mapping.module_slot, [mapping](plugin_block& block) {
@@ -524,6 +524,7 @@ void
 env_engine::reset(plugin_block const* block)
 {
   _stage_pos = 0;
+  _total_pos = 0;
   _current_level = 0;
   _multitrig_level = 0;
   _stage = env_stage::delay;
@@ -572,6 +573,18 @@ env_engine::process(plugin_block& block, cv_cv_matrix_mixdown const* modulation)
     process_mono<false>(block, modulation);
   else
     process_mono<true>(block, modulation);
+
+  // only want the indicators for the actual audio engine
+  if (block.graph) return;
+
+  if (_stage == env_stage::end) return;
+  float flt = block.state.own_block_automation[param_filter][0].real() / 1000.0f;
+  mod_indicator_state indicator_state = {};
+  indicator_state.data.module = module_env;
+  indicator_state.data.module_slot = block.module_slot;
+  indicator_state.data.voice = block.voice->state.slot;
+  indicator_state.data.value = _total_pos / (_dly + _att + _hld + _dcy + _rls + flt);
+  block.push_mod_indicator_state(indicator_state);
 }
 
 template <bool Monophonic>
@@ -705,6 +718,8 @@ void env_engine::process_mono_type_sync_trigger_mode(plugin_block& block, cv_cv_
     {
       _current_level = 0;
       _multitrig_level = 0;
+      _stage_pos += 1.0 / block.sample_rate;
+      _total_pos += 1.0 / block.sample_rate;
       float level = _filter.next(0);
       block.state.own_cv[0][0][f] = level;
       if (level < 1e-5)
@@ -800,6 +815,7 @@ void env_engine::process_mono_type_sync_trigger_mode(plugin_block& block, cv_cv_
     check_unipolar(_multitrig_level);
     block.state.own_cv[0][0][f] = _filter.next(out);
     _stage_pos += 1.0 / block.sample_rate;
+    _total_pos += 1.0 / block.sample_rate;
     if (_stage_pos < stage_seconds) continue;
 
     _stage_pos = 0;