+++ src/firefly_synth/firefly_synth/matrix_osc_osc.cpp
@@ -75,7 +75,8 @@ public:
 };
 
 static graph_data
-render_graph(plugin_state const& state, graph_engine* engine, int param, param_topo_mapping const& mapping)
+render_graph(
+  plugin_state const& state, graph_engine* engine, int param, param_topo_mapping const& mapping)
 {
   int max_osc = 0;
   std::vector<float> result_l;