+++ src/firefly_synth/firefly_synth/voice_in.cpp
@@ -94,7 +94,8 @@ public:
 };
 
 static graph_data
-render_graph(plugin_state const& state, graph_engine* engine, int param, param_topo_mapping const& mapping)
+render_graph(
+  plugin_state const& state, graph_engine* engine, int param, param_topo_mapping const& mapping)
 {
   bool bipolar = mapping.param_index == param_cent;
   std::string partition = state.desc().params[param]->info.name;